chore: organize into subdirs, add gif to readme
