--- compile.bat
@@ -1,1 +1,0 @@
-..\..\..\bin\lcc -o 3d.gb 3d.c primitives.c math.c
